reg-zareg/requirements.txt
fastapi==0.89.1
uvicorn==0.34.0
python-dotenv==1.0.1
SQLAlchemy==1.4.54
psycopg2-binary==2.9.10